SW6: init base code, init migration for new plugin shopware 6
