--- custom/plugins/WnsArShopware/src/Resources/app/administration/src/main.js
@@ -1,1 +1,0 @@
-import './module/wns-sw-arsmundi-background';
