--- custom/plugins/WnsArShopware/src/Resources/app/administration/src/module/wns-sw-arsmundi-background/page/wns-sw-arsmundi-background-list/wns-sw-arsmundi-background-list.html.twig
@@ -1,125 +1,0 @@
-{% block wns_sw_arsmundi_background_list %}
-    <sw-page class="wns-sw-arsmundi-background-list">
-        {% block wwns_sw_arsmundi_background_list_search_bar %}
-            <template slot="search-bar">
-                <sw-search-bar
-                        initial-search-type="background"
-                        :placeholder="$tc('wns-arsmundi.placeholder.search')"
-                        :initial-search="term"
-                        @search="onSearch"
-                />
-            </template>
-        {% endblock %}
-        {% block wns_sw_arsmundi_background_list_smart_bar_header %}
-            <template slot="smart-bar-header">
-                {% block wns_sw_arsmundi_background_list_smart_bar_header_title %}
-                    <h2>
-                        {% block wns_sw_arsmundi_background_list_smart_bar_header_title_text %}
-                            {{ $tc('wns-arsmundi.general.title') }}
-                        {% endblock %}
-
-                        {% block wns_sw_arsmundi_background_list_smart_bar_header_amount %}
-                            <span
-                                    v-if="!isLoading"
-                                    class="sw-page__smart-bar-amount"
-                            >
-                                ({{ total }})
-                            </span>
-                        {% endblock %}
-                    </h2>
-                {% endblock %}
-            </template>
-        {% endblock %}
-
-        {% block wns_sw_arsmundi_background_list_smart_bar_actions %}
-            <template slot="smart-bar-actions">
-                {% block wns_sw_arsmundi_background_list_smart_bar_actions_add %}
-                    <sw-button
-                            class="wns-sw-arsmundi-background-list__button-create"
-                            variant="primary"
-                            :router-link="{ name: 'wns.sw.arsmundi.background.create' }"
-                    >
-                        {{ $tc('wns-arsmundi.button.addButtonText') }}
-                    </sw-button>
-                {% endblock %}
-            </template>
-        {% endblock %}
-        {% block wns_sw_arsmundi_background_list_language_switch %}
-            <template slot="language-switch">
-                <sw-language-switch
-                        slot="language-switch"
-                        @on-change="onChangeLanguage"
-                />
-            </template>
-        {% endblock %}
-
-        <template slot="content">
-            {% block wns_sw_arsmundi_background_list_content %}
-                <div class="wns-sw-arsmundi-background-list__content">
-                    {% block wns_sw_arsmundi_background_list_grid %}
-                        <sw-entity-listing
-                                :repository="backgroundRepository"
-                                :columns="backgroundColumns"
-                                :items="backgrounds"
-                                :is-loading="isLoading"
-                                :sort-by="currentSortBy"
-                                full-page
-                                @page-change="onPageChange"
-                                :allowDelete="true"
-                                :disable-data-fetching="true"
-                                @update-records="updateTotal"
-                                identifier="wns-sw-arsmundi-background-list"
-                        >
-                            {% block wns_sw_arsmundi_background_list_grid_columns_name_preview %}
-                                <template #column-media="{ item }">
-                                    <sw-media-preview-v2 :source="item.media ? item.media.id : null" />
-                                    <router-link
-                                            :to="onEdit(item.id,false)">
-                                        {{ item.media ? item.media.fileName : "No background" }}
-                                    </router-link>
-                                </template>
-                            {% endblock %}
-
-                            {% block sw_product_list_grid_columns_active_content %}
-                                <template #column-active="{ item }">
-                                    <sw-icon
-                                            v-if="item.active"
-                                            name="small-default-checkmark-line-medium"
-                                            small
-                                            color="#37d046"
-                                            class="is--active"
-                                    />
-                                    <sw-icon
-                                            v-else
-                                            name="small-default-x-line-medium"
-                                            small
-                                            color="#de294c"
-                                            class="is--inactive"
-                                    />
-                                </template>
-                            {% endblock %}
-                            {% block sw_product_list_grid_columns_actions %}
-                                <template #more-actions="{ item }">
-                                    {% block sw_product_list_grid_columns_actions_edit %}
-                                        <sw-context-menu-item
-                                                class="wns-sw-arsmundi-background-list__duplicate-action"
-                                                @click="onEdit(item.id)"
-                                        >
-                                            {{ $tc('wns-arsmundi.button.editButtonText') }}
-                                        </sw-context-menu-item>
-                                    {% endblock %}
-                                </template>
-                            {% endblock %}
-                        </sw-entity-listing>
-                        {% block wns_sw_arsmundi_background_list_grid_loader %}
-                            <sw-loader v-if="isLoading" />
-                        {% endblock %}
-                    {% endblock %}
-                </div>
-            {% endblock %}
-        </template>
-
-
-    </sw-page>
-
-{% endblock %}
