--- custom/plugins/WnsArShopware/src/Resources/app/storefront/src/scss/base.scss
@@ -1,3 +1,0 @@
-@import "./vendor/nouislider";
-@import "./components/product-preview";
-@import "./components/product-preview-upload";
