--- custom/plugins/WnsArShopware/src/Resources/app/storefront/src/scss/components/_product-preview-upload.scss
@@ -1,98 +1,0 @@
-// Upload Modal
-.wns-preview-tool-upload-background-modal {
-    .upload-container {
-        text-align: center;
-    }
-
-    .border-container {
-        border: 4px dashed rgba(198, 198, 198, 0.65);
-        text-align: center;
-        padding: 20px;
-        margin-top: 20px;
-
-        .upload-icon {
-            margin-bottom: 10px;
-        }
-
-        /* This sets the actual file input to overlay our button */
-        #upload-file{
-            opacity: 0;
-            -moz-opacity: 0;
-            width:200px;
-            cursor: pointer;
-            position:absolute;
-            left: 50%;
-            transform: translateX(-50%);
-            bottom: 40px;
-            height: 50px;
-        }
-
-        .wns-preview-tool-upload-background-validate-message {
-            color: $sw-color-danger;
-            display: none;
-        }
-    }
-
-    .btn-next-to-setting {
-        display: none;
-    }
-}
-
-// Setting Modal
-.wns-preview-tool-setting-background-image-slider {
-    margin-bottom: 20px;
-
-    &.noUi-target {
-        background: transparent;
-        border-radius: 0;
-        border: none;
-        box-shadow: none;
-    }
-}
-
-.wns-preview-tool-setting-background-image-wrapper {
-    position: relative;
-    background-size: cover;
-
-    > img {
-        width: 100%;
-        height: 360px;
-        max-width: 100%;
-        object-fit: cover;
-    }
-}
-
-.wns-preview-tool-setting-background-image-overlay-left {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    width: 25%;
-    background-color: rgba(0, 0, 0, 0.2);
-    transition: all 0.2s linear;
-}
-
-.wns-preview-tool-setting-background-image-overlay-right {
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    right: 0;
-    width: 25%;
-    background-color: rgba(0, 0, 0, 0.2);
-    transition: all 0.2s linear;
-}
-
-.empty-mess-bg-width-input, .invalid-mess-bg-width-input,
-.empty-mess-bg-top-input, .invalid-mess-bg-top-input,
-.empty-mess-bg-left-input, .invalid-mess-bg-left-input {
-    color: $sw-color-danger;
-    display: none;
-}
-
-.field-required {
-    color: $sw-color-danger;
-}
-
-.bg-top-input, .bg-width-input, .bg-left-input {
-    width: 350px;
-}
