--- custom/plugins/WnsArShopware/src/Resources/app/storefront/src/scss/components/_product-preview.scss
@@ -1,221 +1,0 @@
-
-  .modal-dialog.modal-fullscreen {
-    margin: 0;
-    max-width: 100%;
-    align-items: flex-start;
-
-    .modal-content {
-      padding: 0 20px;
-      height: 100vh;
-    }
-
-    .modal-header {
-      height: 45px;
-    }
-    
-    .modal-body {
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      padding: 0;
-      height: calc(100vh - 40px);
-      iframe{
-        width: 100%;
-        height: 100%;
-        border:0
-      }
-    }
-
-    .close .icon-x {
-      color: black;
-    }
-}
-
-
-.img-thumbnail {
-    border-radius: 0.1875rem;
-    max-width: 13%;
-    height: auto;
-}
-
-
-.wns-product-preview-container{
-  width: 100%;
-  max-width: 100%;
-  overflow: hidden;
-  position: absolute;
-  height: 100vh;
- 
-}
-
-
-.wns-background-dropdown-menu{
-  height: auto;
-  width: 250px;
-  max-height: 400px;
-  overflow-x: hidden;
-  overflow-y: scroll;
-  background-color: #fff;
-  box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-  padding: 15px 0px;
-  .wns-background-dropdown-item {
-    padding: 0;
-    margin-bottom: 2px;
-    max-width: 50%;
-    :hover {
-      background-color: #fff;
-    }
-    .img-hover-zoom {
-      height: 120px; 
-      overflow: hidden; 
-    }
-    .img-hover-zoom img {
-      transition: transform .5s ease;
-      max-width: 100%;
-      width: 100%;
-      height: 120px;
-      object-fit: cover;
-    }
-
-    .img-hover-zoom:hover img {
-      transform: scale(1.1);
-      box-shadow: rgba(0, 0, 0, 0.24) 0px 2px 6px;
-    }
-  }
-}
-
-.wns-product-preview-action-sidebar {
-  position: absolute;
-  height: 60%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-left: 20px;
-  .wns-product-preview-action-button-group{
-    display: flex;
-    flex-direction: column;
-    gap: 5px;
-    .btn-secondary{
-      height: 50px;
-      width: 130px;
-    }
-    .wns-product-preview-action-share-product-preview .dropdown-toggle::after {
-      display: none;
-  }
-  }
-}
-
-
-.wns-product-preview-background {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  background: no-repeat center;
-  background-size: 100% auto;
-  transition: background-image 1s ease-in-out;
-}
-
-
-.wns-product-preview-product-image-container{
-  position: absolute;
-  height: 100%;
-  width: 100%;
-}
-
-.wns-product-preview-product-image {
-  z-index: 9000;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-
-  .wns-product-preview-dragged-image {
-    position: absolute;
-    transition: width 1s ease-in-out;
-    :hover {
-      cursor: move;
-    }
-    .moving {
-      transition: top 1s ease, left 1s ease;
-    }
-    
-  }
-  img {
-    display: block;
-    max-width: 50%;
-    height: auto;
-    box-shadow: rgba(0, 0, 0, 0.09) 0px 2px 1px, rgba(0, 0, 0, 0.09) 0px 4px 2px, rgba(0, 0, 0, 0.09) 0px 8px 4px, rgba(0, 0, 0, 0.09) 0px 16px 8px, rgba(0, 0, 0, 0.09) 0px 32px 16px;
-  }
-}
-
-
-.wns-preview-tool-action-share-link{
-  position: relative;
-  top:15px;
-  left:0px;
-  z-index: 999;
-  background: rgb(255, 255, 255);
-  border-radius: 4px;
-  width: 450px;
-  box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-  padding: 15px 15px;
-  .wns-preview-tool-share-link-input-group {
-    position: relative;
-    background: #fff;
-    border: 1px solid rgb(192, 191, 191);
-    border-radius: 3px;
-    display: flex;
-    input{
-      background-color: #fff;
-      color: #555;
-      border: none;
-      outline: none;
-    }
-    .wns-copy-btn{
-      padding: 2px 10px;
-      background: #fff;
-      border: none;
-      outline: none;
-      border-radius: 10px;
-      cursor: pointer;
-    }
-    .wns-copy-icon .icon.icon-products{
-      color: #008490;
-    }
-    button:hover .wns-copy-icon .icon.icon-products{
-      color: #08565d;
-    }
-    
-    .tooltiptext {
-      width: 140px;
-      background-color: #555;
-      color: rgb(248, 245, 245);
-      text-align: center;
-      border-radius: 6px;
-      padding: 5px;
-      position: absolute;
-      z-index: 99999999;
-      bottom: 120%;
-      left: 78%;
-      opacity: 0;
-      transition: opacity 0.3s;
-    }
-    button:hover .tooltiptext {
-      visibility: visible;
-      opacity: 0.8;
-    }
-    button:not( :hover ) .tooltiptext{
-      opacity: 0;
-    }
-    .tooltiptext::after {
-      content: "";
-      position: absolute;
-      top: 100%;
-      left: 50%;
-      margin-left: -5px;
-      border-width: 5px;
-      border-style: solid;
-      border-color: #555 transparent transparent transparent;
-    }
-  }
-   
-}
