+++ custom/plugins/WnsArShopware/src/Resources/config/routes.xml
@@ -5,4 +5,5 @@
         https://symfony.com/schema/routing/routing-1.0.xsd">
 
     <import resource="../../Storefront/**/Controller/*Controller.php" type="annotation" />
+    <import resource="../../Storefront/**/Controller/TestPlugin/*Controller.php" type="annotation" />
 </routes>
