+++ custom/plugins/WnsArShopware/src/Resources/config/services.xml
@@ -13,36 +13,13 @@
             <tag name="shopware.entity.definition" entity="wns_image_background_upload_sales_channel" />
         </service>
 
-        <service id="Wns\WnsArShopware\Command\ImageBackgroundUploadDemoDataCommand">
-            <argument type="service" id="wns_image_background_upload.repository"/>
-            <argument type="service" id="wns_image_background_upload_sales_channel.repository"/>
-            <tag name="console.command" />
-        </service>
-
-        <service id="Wns\WnsArShopware\Storefront\Controller\ImageBackgroundController" public="true">
-            <argument type="service" id="media_folder.repository" />
-            <argument type="service" id="media.repository" />
+        <service id="Wns\WnsArShopware\Storefront\Controller\TestPlugin\TestPluginController" public="true">
             <argument type="service" id="wns_image_background_upload.repository" />
             <argument type="service" id="wns_image_background_upload_sales_channel.repository" />
-            <argument type="service" id="Shopware\Core\Content\Media\File\FileSaver" />
-            <argument type="service" id="Shopware\Core\Content\Media\File\FileNameProvider" />
-            <argument type="service" id="Shopware\Core\System\SystemConfig\SystemConfigService" />
-            <argument type="service" id="Wns\WnsArShopware\Storefront\Page\ImageBackground\ImageBackgroundPageLoader" />
+
             <call method="setContainer">
-                <argument type="service" id="service_container" />
+                <argument type="service" id="service_container"/>
             </call>
-        </service>
-
-        <service id="Wns\WnsArShopware\Subscriber\CustomCoreSubscriber">
-            <tag name="kernel.event_subscriber"/>
-        </service>
-
-        <service id="Wns\WnsArShopware\Storefront\Page\ImageBackground\ImageBackgroundPageLoader" public="true">
-            <argument type="service" id="Shopware\Storefront\Page\GenericPageLoader" />
-            <argument type="service" id="wns_image_background_upload.repository" />
-            <argument type="service" id="Shopware\Core\System\SystemConfig\SystemConfigService" />
-            <argument type="service" id="mail_template_type.repository"/>
-            <argument type="service" id="Shopware\Core\Framework\Adapter\Twig\StringTemplateRenderer"/>
         </service>
     </services>
 </container>
