--- custom/plugins/WnsArShopware/src/Resources/snippet/de_DE/message.de-DE.json
@@ -1,5 +1,0 @@
-{
-   "uploadSuccess": "Ihr Hintergrundbild wurde erfolgreich hochgeladen.",
-    "uploadFail": "Dieses Bild kann nicht hochgeladen werden, bitte versuchen Sie es später erneut.",
-    "invalidFile": "Hochgeladene Bilddateigröße ist zu groß oder falsche Erweiterung."
-}
