--- custom/plugins/WnsArShopware/src/Resources/snippet/en_GB/message.en-GB.json
@@ -1,5 +1,0 @@
-{
-    "uploadSuccess": "Your background image is uploaded successfully.",
-    "uploadFail": "Can not upload this image, please try again later.",
-    "invalidFile": "Uploaded image file size is too large or wrong extension."
-}
