--- custom/plugins/WnsArShopware/src/Resources/views/storefront/page/product-detail/index.html.twig
@@ -1,53 +1,0 @@
-{% sw_extends '@Storefront/storefront/page/product-detail/index.html.twig' %}
-
-{% block page_product_detail_media %}
-    <div class="col-lg-7 product-detail-media">
-        {% if page.product.media %}
-            {% sw_include '@Storefront/storefront/element/cms-element-image-gallery.html.twig' with {
-                'mediaItems': mediaItems,
-                'zoom': true,
-                'zoomModal': true,
-                'displayMode': 'contain',
-                'gutter': 5,
-                'minHeight': '430px',
-                'navigationArrows': 'inside',
-                'navigationDots': 'inside',
-                'galleryPosition': 'left',
-                'isProduct': true,
-                'fallbackImageTitle': page.product.translated.name,
-                'startIndexThumbnails': page.product.cover.position + 1,
-                'startIndexSlider': page.product.cover.position + 1,
-                'keepAspectRatioOnZoom': false
-            } %}
-        {% endif %}
-        {% block wns_product_preview_modal %}
-            <img src="{{ asset('bundles/wnsarshopware/button-preview-product.jpg', 'asset') }}"
-                 data-toggle="modal" data-target="#previewProductModal" class="img-thumbnail img-open-tool"
-                 alt="">
-
-            <!-- Modal -->
-            <div class="modal fade" id="previewProductModal" tabindex="-1" role="dialog" aria-labelledby="previewProductModalLabel" aria-hidden="true">
-                <div class="modal-dialog modal-fullscreen">
-                    <div class="modal-content">
-                        <div class="modal-header">
-                            <h5 class="modal-title" id="previewProductModalLabel">{{page.product.translated.name }}</h5>
-                            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
-                                {% sw_icon 'x' style {
-                                    'size': 'sm'
-                                } %}
-                            </button>
-                        </div>
-                    {% block wns_product_preview_modal_body %}
-                        <div class="modal-body">
-                            <iframe
-                                src="{{ url('frontend.product.preview',{product_id: page.product.id, productimg_url: page.product.cover.media.url, productimg_width: page.product.width ? page.product.width/10 : 0,  productimg_height: page.product.height ? page.product.height/10 : 0,  productimg_id: page.product.id }) }}">
-                            </iframe>
-                        </div>
-                    {% endblock %}
-
-                    </div>
-                </div>
-            </div>
-        {% endblock %}
-    </div>
-{% endblock %}
